Address some Code Climate suggestions
