+++ lib/numbers_and_words/strategies/figures_converter/options/base.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'numbers_and_words/strategies/figures_converter/options/base/gender'
+require 'numbers_and_words/strategies/figures_converter/options/base/ordinal'
+require 'numbers_and_words/strategies/figures_converter/options/base/remove_zero'
