+++ lib/numbers_and_words/strategies/figures_converter/options/cs/gender.rb
@@ -5,22 +5,9 @@ module NumbersAndWords
     module FiguresConverter
       module Options
         module Cs
-          class Gender
-            attr_accessor :strategy, :options
-
-            def initialize(proxy, *_args)
-              @strategy = proxy.strategy
-              @options = proxy.options
-            end
-
+          class Gender < Base::Gender
             def result
               active? ? @options[:gender].to_sym : :female
-            end
-
-            private
-
-            def active?
-              @options[:gender]
             end
           end
         end
