+++ lib/numbers_and_words/strategies/figures_converter/options/cs/ordinal.rb
@@ -5,21 +5,7 @@ module NumbersAndWords
     module FiguresConverter
       module Options
         module Cs
-          class Ordinal
-            def initialize(proxy, *_args)
-              @strategy = proxy.strategy
-              @options = proxy.options
-            end
-
-            def result
-              return :ordinal if active?
-            end
-
-            private
-
-            def active?
-              @options[:ordinal]
-            end
+          class Ordinal < Base::Ordinal
           end
         end
       end
