+++ lib/numbers_and_words/strategies/figures_converter/options/es/remove_zero.rb
@@ -5,23 +5,7 @@ module NumbersAndWords
     module FiguresConverter
       module Options
         module Es
-          class RemoveZero
-            attr_accessor :strategy, :options
-
-            def initialize(proxy, *_args)
-              @strategy = proxy.strategy
-              @options = proxy.options
-            end
-
-            def result
-              active?
-            end
-
-            private
-
-            def active?
-              @options[:remove_zero]
-            end
+          class RemoveZero < Base::RemoveZero
           end
         end
       end
