+++ lib/numbers_and_words/strategies/figures_converter/options/ru/gender.rb
@@ -5,23 +5,7 @@ module NumbersAndWords
     module FiguresConverter
       module Options
         module Ru
-          class Gender
-            attr_accessor :strategy, :options
-
-            def initialize(proxy, *_args)
-              @strategy = proxy.strategy
-              @options = proxy.options
-            end
-
-            def result
-              active? ? @options[:gender].to_sym : :male
-            end
-
-            private
-
-            def active?
-              @options[:gender]
-            end
+          class Gender < Base::Gender
           end
         end
       end
